created error message component, issue in story
