+++ apps/gym/src/components/reusable/DateRangePicker/DateRangePicker.module.css
@@ -12,10 +12,3 @@
   flex-flow: row wrap;
   gap: 5px
 }
-
-.error_message {
-  color: var(--bright-red);
-  font-size: 20px;
-  height: min-content;
-  margin: auto 5px;
-}
